Added new compressor class. Counters class not has a lot of this keywords and no recursive errors. Put @Override on LimitSwitch function and replased some this keywords with super. Deleted MoveByTime commands for DriveForTime command. Lots of commenting
