+++ src/org/usfirst/frc/falcons6443/robot/commands/TeleopMode.java
@@ -1,11 +1,7 @@
 package org.usfirst.frc.falcons6443.robot.commands;
 
-import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import org.usfirst.frc.falcons6443.robot.Robot;
-import org.usfirst.frc.falcons6443.robot.hardware.Joysticks.Xbox;
-import org.usfirst.frc.falcons6443.robot.utilities.Logger;
-import org.usfirst.frc.falcons6443.robot.utilities.enums.*;
+import org.usfirst.frc.falcons6443.robot.hardware.joysticks.Xbox;
 
 import java.util.List;
 import java.util.ArrayList;
