+++ src/org/usfirst/frc/falcons6443/robot/hardware/LimitSwitch.java
@@ -2,6 +2,10 @@ package org.usfirst.frc.falcons6443.robot.hardware;
 
 import edu.wpi.first.wpilibj.DigitalInput;
 
+/*
+ * Creates a limit switch, inverting the output so that get() returns true when the
+ * limit switch is pushed and false if it is not pushed.
+ */
 public class LimitSwitch extends DigitalInput {
     private boolean uninverted = false;
 
@@ -14,10 +18,11 @@ public class LimitSwitch extends DigitalInput {
         this.uninverted = uninverted;
     }
 
+    @Override
     //returns true if pushed, false if not pushed (false if pushed when uninverted is true)
     public boolean get() {
-        if (!uninverted) return !this.get();
-        else return this.get();
+        if (!this.uninverted) return !super.get();
+        else return super.get();
     }
 
     public void setUninverted(boolean uninverted) { this.uninverted = uninverted; }
