+++ src/org/usfirst/frc/falcons6443/robot/hardware/joysticks/Playstation.java
@@ -1,19 +1,23 @@
-package org.usfirst.frc.falcons6443.robot.hardware.Joysticks;
+package org.usfirst.frc.falcons6443.robot.hardware.joysticks;
 
 import edu.wpi.first.wpilibj.Joystick;
 
-//NOTE: no rumble
-public class Logitech {
+/**
+ * Wrapper for a playstation controller. Provides clearer interface with button and axis inputs.
+ * Use class in Teleop and OI classes
+ **/
+public class Playstation {
 
     public Joystick controller;
 
-    public Logitech(Joystick controller) {
+    public Playstation(Joystick controller) {
         this.controller = controller;
     }
 
-    /**
-     * @return The value of the X axis of the left stick.
-     */
+    public Joystick getJoystick(Joystick play) {
+        return play;
+    }
+
     public double leftStickX() {
         return controller.getRawAxis(0);
     }
@@ -29,28 +33,28 @@ public class Logitech {
      * @return The value of the X axis of the right stick.
      */
     public double rightStickX() {
-        return controller.getRawAxis(4);
+        return controller.getRawAxis(2);
     }
 
     /**
      * @return The value of the Y axis of the right stick.
      */
     public double rightStickY() {
-        return controller.getRawAxis(5);
+        return controller.getRawAxis(3);
     }
 
     /**
      * @return The value of the axis for the left trigger.
      */
     public boolean leftTrigger() {
-        return controller.getRawButton(2);
+        return controller.getRawButton(7);
     }
 
     /**
      * @return The value of the axis for the right trigger.
      */
     public boolean rightTrigger() {
-        return controller.getRawButton(3);
+        return controller.getRawButton(8);
     }
 
     /**
@@ -71,21 +75,21 @@ public class Logitech {
      * @return the value of the left joystick button.
      */
     public boolean leftStickButton() {
-        return controller.getRawButton(9);
+        return controller.getRawButton(11);
     }
 
     /**
      * @return the value of the right joystick button.
      */
     public boolean rightStickButton() {
-        return controller.getRawButton(10);
+        return controller.getRawButton(12);
     }
 
     /**
      * @return the value of the A button.
      */
     public boolean A() {
-        return controller.getRawButton(1);
+        return controller.getRawButton(3);
     }
 
     /**
@@ -99,25 +103,19 @@ public class Logitech {
      * @return the value of the X button.
      */
     public boolean X() {
-        return controller.getRawButton(3);
+        return controller.getRawButton(4);
     }
 
     /**
      * @return the value of the Y button.
      */
     public boolean Y() {
-        return controller.getRawButton(4);
+        return controller.getRawButton(1);
     }
 
-    /**
-     * @return the value of the back button.
-     */
-    public boolean back() {
-        return controller.getRawButton(7);
+    //PLEASE FIND THE ACTUAL NAME FOR THESE TOO
+    public boolean seven() {
+        return controller.getRawButton(9);
     }
-
-    /**
-     * @return the value of the start button.
-     */
-    public boolean start() { return controller.getRawButton(8); }
+    public boolean eight() { return controller.getRawButton(10); }
 }
